Changed coins to int rather than double

+++ src/main/java/seng201/team0/models/Purchasable.java
@@ -1,5 +1,5 @@
 package seng201.team0.models;
 
 public interface Purchasable {
-    public double getCost();
+    public int getCost();
 }
